Differential Gene Expression Analysis by Limma

utility function to get differential expression table from limma function in limma package.

limma_deg(gex, label)

Arguments

gex
A gene x sample data frame. rownames are gene or probe ids and colnames are sample ids.

label
A factor. Indicates class label of samples. It should be ordered as sample ids in gex data frame.

Value

DEG data frame which contains logFC, AveExpr, t, P.Value, adj.P.Val (by Benjamin-Hochberg correction)

Author

Seongyong Park (2020-08-18)

Examples

gex = do.call(rbind, lapply(1:100, function(i) rnorm(100, sample(1:10), 2)))
rownames(gex) = paste0("gid", 1:100)
colnames(gex) = paste0("sid", 1:100)
label = factor(c(rep("ALL", 50), rep("AML", 50)), levels=c("ALL", "AML"))
limma_deg(gex, label)

#>                logFC  AveExpr            t   P.Value adj.P.Val         B
#> gid34   0.8193437298 5.657343  1.175511760 0.2398186 0.9991119 -4.591372
#> gid48  -0.7666263025 5.870307 -1.099878111 0.2714122 0.9991119 -4.593108
#> gid12  -0.7661176389 5.466124 -1.099148332 0.2717304 0.9991119 -4.593124
#> gid14  -0.7638359190 5.772627 -1.095874750 0.2731605 0.9991119 -4.593197
#> gid58  -0.7562947524 5.877817 -1.085055445 0.2779238 0.9991119 -4.593435
#> gid90   0.7189914446 5.168472  1.031536421 0.3023148 0.9991119 -4.594577
#> gid61  -0.6973683005 5.871145 -1.000513714 0.3170867 0.9991119 -4.595213
#> gid94   0.6959858970 5.438175  0.998530381 0.3180469 0.9991119 -4.595253
#> gid86   0.6885739685 5.624040  0.987896493 0.3232277 0.9991119 -4.595466
#> gid32  -0.6521900729 5.698548 -0.935696549 0.3494525 0.9991119 -4.596478
#> gid100  0.6115421335 5.115076  0.877378984 0.3803023 0.9991119 -4.597544
#> gid13  -0.6054296784 5.280439 -0.868609450 0.3850821 0.9991119 -4.597699
#> gid83   0.6012614580 5.161356  0.862629308 0.3883624 0.9991119 -4.597803
#> gid73   0.5983077366 5.619104  0.858391606 0.3906973 0.9991119 -4.597877
#> gid81   0.5949677866 5.900199  0.853599783 0.3933476 0.9991119 -4.597959
#> gid52   0.5914220025 5.554548  0.848512649 0.3961733 0.9991119 -4.598047
#> gid62  -0.5804780393 5.656618 -0.832811355 0.4049714 0.9991119 -4.598313
#> gid80  -0.5702767809 5.382141 -0.818175618 0.4132768 0.9991119 -4.598557
#> gid22  -0.5678441874 5.604801 -0.814685578 0.4152721 0.9991119 -4.598614
#> gid26  -0.5516955863 5.564649 -0.791517194 0.4286614 0.9991119 -4.598989
#> gid53   0.5468376010 5.638351  0.784547446 0.4327379 0.9991119 -4.599100
#> gid38   0.5434170565 5.633710  0.779639994 0.4356216 0.9991119 -4.599178
#> gid70   0.5404227953 5.638916  0.775344130 0.4381550 0.9991119 -4.599245
#> gid3    0.5167105499 5.413189  0.741324191 0.4585147 0.9991119 -4.599765
#> gid91  -0.4719342754 5.556635 -0.677083708 0.4983688 0.9991119 -4.600684
#> gid76   0.4671124470 5.179674  0.670165835 0.5027679 0.9991119 -4.600778
#> gid75   0.4633515412 5.235697  0.664770066 0.5062132 0.9991119 -4.600851
#> gid54   0.4631502000 5.268777  0.664481202 0.5063980 0.9991119 -4.600855
#> gid35   0.4582362968 5.548670  0.657431229 0.5109191 0.9991119 -4.600949
#> gid65  -0.4538325873 5.431329 -0.651113231 0.5149887 0.9991119 -4.601032
#> gid29  -0.4494710458 5.560349 -0.644855731 0.5190358 0.9991119 -4.601114
#> gid18  -0.4132649060 5.594218 -0.592910813 0.5532545 0.9991119 -4.601762
#> gid79   0.4066231479 5.552697  0.583381888 0.5596497 0.9991119 -4.601875
#> gid69  -0.4051428870 5.284989 -0.581258159 0.5610798 0.9991119 -4.601900
#> gid59  -0.4004686473 5.367339 -0.574552031 0.5656075 0.9991119 -4.601978
#> gid16   0.3936288857 5.689119  0.564739031 0.5722642 0.9991119 -4.602091
#> gid96  -0.3862843784 5.727081 -0.554201873 0.5794533 0.9991119 -4.602210
#> gid39  -0.3787261268 5.567310 -0.543358055 0.5868957 0.9991119 -4.602330
#> gid85  -0.3777669087 5.586389 -0.541981866 0.5878433 0.9991119 -4.602345
#> gid9    0.3650305802 5.735740  0.523709066 0.6004928 0.9991119 -4.602541
#> gid66  -0.3635177060 5.102472 -0.521538547 0.6020034 0.9991119 -4.602564
#> gid95  -0.3536730963 4.969061 -0.507414494 0.6118754 0.9991119 -4.602711
#> gid25  -0.3531581208 5.451026 -0.506675659 0.6123938 0.9991119 -4.602718
#> gid31   0.3450658353 5.345018  0.495065664 0.6205649 0.9991119 -4.602836
#> gid63   0.3435314492 5.451070  0.492864281 0.6221195 0.9991119 -4.602858
#> gid82  -0.3274914798 5.645555 -0.469851751 0.6384714 0.9991119 -4.603081
#> gid10  -0.3254792805 5.304797 -0.466964850 0.6405354 0.9991119 -4.603108
#> gid4   -0.3241353688 5.632606 -0.465036741 0.6419155 0.9991119 -4.603126
#> gid57   0.3209212556 5.732344  0.460425455 0.6452211 0.9991119 -4.603169
#> gid71   0.3061883921 5.292059  0.439288229 0.6604624 0.9991119 -4.603361
#> gid44   0.2939334695 5.500353  0.421706101 0.6732488 0.9991119 -4.603514
#> gid60   0.2827994566 5.394903  0.405732142 0.6849483 0.9991119 -4.603647
#> gid36  -0.2741623000 5.363534 -0.393340421 0.6940766 0.9991119 -4.603747
#> gid40   0.2550717688 5.550897  0.365951252 0.7144094 0.9991119 -4.603957
#> gid51   0.2508653765 5.660990  0.359916345 0.7189174 0.9991119 -4.604001
#> gid8   -0.2472306393 5.722843 -0.354701590 0.7228208 0.9991119 -4.604038
#> gid67  -0.2404309198 5.213805 -0.344946038 0.7301423 0.9991119 -4.604107
#> gid88  -0.2268349370 5.496180 -0.325439893 0.7448551 0.9991119 -4.604239
#> gid15   0.2222837202 5.612572  0.318910266 0.7498014 0.9991119 -4.604282
#> gid84  -0.2188336917 5.160948 -0.313960513 0.7535577 0.9991119 -4.604313
#> gid27  -0.2181986866 5.452439 -0.313049472 0.7542497 0.9991119 -4.604319
#> gid87  -0.2107437055 5.522917 -0.302353817 0.7623888 0.9991119 -4.604385
#> gid99  -0.2059164264 5.833707 -0.295428124 0.7676731 0.9991119 -4.604427
#> gid43   0.2019917849 5.503036  0.289797444 0.7719773 0.9991119 -4.604460
#> gid89  -0.1987266449 5.797304 -0.285112950 0.7755636 0.9991119 -4.604487
#> gid68  -0.1959553364 5.663336 -0.281136956 0.7786113 0.9991119 -4.604510
#> gid20  -0.1947929461 5.641437 -0.279469276 0.7798906 0.9991119 -4.604520
#> gid41   0.1918229815 5.355521  0.275208270 0.7831620 0.9991119 -4.604543
#> gid33  -0.1908458092 5.958952 -0.273806322 0.7842392 0.9991119 -4.604551
#> gid30   0.1870271119 5.657480  0.268327640 0.7884528 0.9991119 -4.604581
#> gid50   0.1833647106 5.774832  0.263073196 0.7924997 0.9991119 -4.604609
#> gid97   0.1686678354 5.726559  0.241987601 0.8087948 0.9991119 -4.604717
#> gid11   0.1669535108 5.446911  0.239528062 0.8107011 0.9991119 -4.604729
#> gid47  -0.1647110681 5.404004 -0.236310831 0.8131964 0.9991119 -4.604744
#> gid21   0.1616899199 5.475119  0.231976392 0.8165612 0.9991119 -4.604765
#> gid17  -0.1452022749 5.532865 -0.208321582 0.8349822 0.9991119 -4.604870
#> gid37   0.1433847139 5.519722  0.205713928 0.8370187 0.9991119 -4.604880
#> gid6   -0.1349587474 5.158158 -0.193625201 0.8464734 0.9991119 -4.604929
#> gid1    0.1317586473 5.587974  0.189034020 0.8500701 0.9991119 -4.604947
#> gid28  -0.1192915400 5.881596 -0.171147472 0.8641113 0.9991119 -4.605012
#> gid2    0.1162071270 5.367242  0.166722267 0.8675920 0.9991119 -4.605027
#> gid98  -0.1133493795 5.425546 -0.162622259 0.8708192 0.9991119 -4.605040
#> gid49   0.1034015670 5.377576  0.148350141 0.8820695 0.9991119 -4.605085
#> gid72   0.1008183065 5.505486  0.144643939 0.8849950 0.9991119 -4.605096
#> gid5    0.0812662015 5.273937  0.116592551 0.9071853 0.9991119 -4.605170
#> gid64   0.0727531794 5.450588  0.104378925 0.9168708 0.9991119 -4.605197
#> gid56  -0.0685962750 5.336793 -0.098415018 0.9216048 0.9991119 -4.605209
#> gid24   0.0651185955 5.614044  0.093425594 0.9255674 0.9991119 -4.605219
#> gid19   0.0591355423 5.419885  0.084841713 0.9323890 0.9991119 -4.605235
#> gid78   0.0465253716 5.318946  0.066749911 0.9467822 0.9991119 -4.605262
#> gid74  -0.0450894501 5.252693 -0.064689796 0.9484223 0.9991119 -4.605265
#> gid93   0.0448003179 5.374905  0.064274978 0.9487526 0.9991119 -4.605266
#> gid7    0.0432794486 5.331482  0.062092989 0.9504900 0.9991119 -4.605268
#> gid77  -0.0312428659 5.516939 -0.044824113 0.9642484 0.9991119 -4.605287
#> gid46   0.0264438721 5.751087  0.037939001 0.9697371 0.9991119 -4.605293
#> gid42   0.0248052690 5.386749  0.035588099 0.9716115 0.9991119 -4.605294
#> gid55  -0.0217105551 5.458961 -0.031148115 0.9751521 0.9991119 -4.605297
#> gid23   0.0120892726 5.405804  0.017344469 0.9861622 0.9991119 -4.605304
#> gid92   0.0018235546 5.535399  0.002616252 0.9979126 0.9991119 -4.605307
#> gid45   0.0007758602 5.526725  0.001113126 0.9991119 0.9991119 -4.605307